chore: remove the row-based measure query planner
The vec dispatch had exactly one fall-through to the row path, the
--measure-vectorized-enabled=false rollback rail, so removing the flag
makes the row planner unreachable.

Drops VectorizedConfig.Enabled, collapses Dispatch's handled return and
processor.go's tryVecDispatch into an iterator-or-error contract, and
deletes the row measure plan nodes. sortedMIterator, sortableDataPoints
and hashDataPoint move out of measure_plan_distributed.go verbatim: the
vec multi-group path still merges through them via MergeGroupMIterators.
BuildSchema moves into schema.go and loses its Deprecated marker, since
it is the only measure schema builder. The TopN pre-aggregation stack is
untouched.

--measure-vectorized-enabled stays registered as a no-op that accepts
=true and fails fast on =false naming apache/skywalking#13998, so an
operator's surviving rollback command line reports the removal instead
of "unknown flag". It goes away in 0.13.0.
