+++ banyand/measure/svc_liaison.go
@@ -153,14 +153,10 @@ func (s *liaison) PreRun(ctx context.Context) error {
 	// Publish the per-process wire mode for TopicInternalMeasureQuery.
 	// On the liaison this controls the DECODE side: an incoming response
 	// body on TopicInternalMeasureQuery is parsed as raw frame bytes
-	// under flag-on (G9f.5.c's distributedPlan recognizes the []byte
-	// shape and runs ReduceFramesToInternalDataPoints /
-	// DecodeFramesToInternalDataPoints) or proto under flag-off. The
-	// per-process modes MUST match across the cluster — the runbook
-	// covers detection + recovery for a partial-rollout skew.
-	data.SetMeasureWireModeRaw(s.option.vectorized.Enabled)
+	// (G9f.5.c's distributed plan recognizes the []byte shape and runs
+	// ReduceFramesToInternalDataPoints / DecodeFramesToInternalDataPoints).
+	data.SetMeasureWireModeRaw(true)
 	s.l.Info().
-		Bool("measure_vectorized_enabled", s.option.vectorized.Enabled).
 		Bool("measure_wire_mode_raw", data.MeasureWireModeRaw()).
 		Msg("G9f wire mode published for TopicInternalMeasureQuery (liaison)")
 	s.lfs = fs.NewLocalFileSystemWithLoggerAndLimit(s.l, s.pm.GetLimit())
